GameState comments for clarity, GeneralCommand adjusted for end turn stuff

+++ ITSD-DT2021-Template/app/commands/GeneralCommandSets.java
@@ -177,12 +177,12 @@ public class GeneralCommandSets {
 	// Show entire Hand 
 	public static void drawCardsInHand(ActorRef out, GameState gameState, int oldHandSize, ArrayList<Card> cardsInHand) {
 
-		// Delete/hide all cards in the UI
-		for (int i = 0; i < oldHandSize; i++) {
-			BasicCommands.deleteCard(out, i);
+		if(gameState.isTwoPlayerMode() ) {// Delete/hide all cards in the UI
+			for (int i = 0; i < oldHandSize; i++) {
+				BasicCommands.deleteCard(out, i);
+			}
+			GeneralCommandSets.threadSleep(); 
 		}
-		GeneralCommandSets.threadSleep(); 
-
 		 
 		// Show all the cards in the UI in new positions 
 		int i = 0;
